Check user's MEME install

In order to use the run* family of functions, memes must detect a local install of the MEME Suite. MEME is installed in a directory named meme/bin/ which can be located anywhere on the filesystem, but is typically found in ~/meme/bin. If the MEME Suite is installed at ~/meme/bin, memes can autodetect the install. However, in the case that the MEME Suite is found at a nonstandard location, the user may specify the location of their meme/bin in three ways:

check_meme_install(meme_path = NULL)

Arguments

meme_path
path to "meme/bin" (if unset will search MEME_BIN environment variable or meme_bin option)

Value

message indicating which MEME utilities are installed and their location on disk

Details

provide the full path to meme/bin to the meme_path argument to each run* function.

set the meme_bin option using options(meme_bin = "path/to/meme/bin") once per R session.

set the MEME_BIN environment variable either in .Renviron or ~/.bashrc with the path to meme/bin

To aid the user in determining if memes can detect their meme/bin install, check_meme_install() will search the aforementioned locations for a valid meme/bin, returning green checks for each detected tool, or red X's for undetected tools. Alternatively, users can run meme_is_installed() to get a boolean value indicating whether their MEME Suite can be detected.

check_meme_install() searches using the following heirarchy. This heirarchy mimics how all run* functions search for meme/bin, thus the paths printed from check_meme_install() will indicate the paths used by each run* function. The heirarchy is as follows:

the meme_path function argument if set

the meme_bin option

the MEME_BIN environment variable

the default location at ~/meme/bin

Examples

check_meme_install()
#> checking main install
#> ✔ /opt/meme/bin
#> checking util installs
#> ✔ /opt/meme/bin/dreme
#> ✔ /opt/meme/bin/ame
#> ✔ /opt/meme/bin/fimo
#> ✔ /opt/meme/bin/tomtom
#> ✔ /opt/meme/bin/meme
#> ✔ /opt/meme/bin/streme