Drop best match columns from tomtom results

Convenience function for dropping all columns created by runTomTom prefixed by "best_match_" and the "best_db_name" column. Keeps the "tomtom" data.frame column. Can be useful if you want to unnest the `tomtom` data without propagating these columns.

drop_best_match(res)

Arguments

res
results of runTomTom

Value

`res` without the tomtom best_match_ columns

Examples

data("example_dreme_tomtom")
names(example_dreme_tomtom)

#>  [1] "motif"              "name"               "altname"           
#>  [4] "family"             "organism"           "consensus"         
#>  [7] "alphabet"           "strand"             "icscore"           
#> [10] "nsites"             "bkgsites"           "pval"              
#> [13] "qval"               "eval"               "type"              
#> [16] "bkg"                "rank"               "seq"               
#> [19] "length"             "positive_hits"      "negative_hits"     
#> [22] "unerased_evalue"    "positive_total"     "negative_total"    
#> [25] "pos_frac"           "neg_frac"           "best_match_name"   
#> [28] "best_match_altname" "best_db_name"       "best_match_offset" 
#> [31] "best_match_pval"    "best_match_eval"    "best_match_qval"   
#> [34] "best_match_strand"  "best_match_motif"   "tomtom"            

names(drop_best_match(example_dreme_tomtom))

#>  [1] "motif"           "name"            "altname"         "family"         
#>  [5] "organism"        "consensus"       "alphabet"        "strand"         
#>  [9] "icscore"         "nsites"          "bkgsites"        "pval"           
#> [13] "qval"            "eval"            "type"            "bkg"            
#> [17] "rank"            "seq"             "length"          "positive_hits"  
#> [21] "negative_hits"   "unerased_evalue" "positive_total"  "negative_total" 
#> [25] "pos_frac"        "neg_frac"        "tomtom"