Parse AME output

This imports AME results using the "ame.tsv" output, and optionally the "sequences.tsv" output if run with "method = fisher". AME results differ based on the method used, thus this must be set on import or the column names will be incorrect.

importAme(path, method = "fisher", sequences = FALSE)

Arguments

path
path to ame results file ("ame.tsv")

method
ame run method used (one of: c("fisher", "ranksum", "dmhg3", "dmhg4", "pearson", "spearman")). Default: "fisher".

sequences
FALSE or path to sequences file (only valid for method = "fisher")

Value

data.frame with method-specific results. See [AME results](http://meme-suite.org/doc/ame-output-format.html) webpage for more information. If sequences is set to a path to the sequences.tsv and method = "fisher", the list-column `sequences` will be appended to resulting data.frame.

See also

Examples

ame_tsv <- system.file("extdata", "ame.tsv", package = "memes", mustWork = TRUE)
importAme(ame_tsv)
#> # A tibble: 107 × 17
#>     rank motif_db   motif_id motif_alt_id consensus   pvalue adj.pvalue   evalue
#>    <int> <chr>      <chr>    <chr>        <chr>        <dbl>      <dbl>    <dbl>
#>  1     1 /pine/scr… Aef1_SO… FBgn0005694… KGTTGTTG… 7.61e-62   3.1 e-59 1.88e-56
#>  2     2 /pine/scr… CG4360-… FBgn0038787… KTTGTTGT… 3.57e-56   1.29e-53 7.87e-51
#>  3     3 /pine/scr… Trl_Fly… FBgn0013263  YYKCTCTC… 2.37e-53   2.48e-50 1.51e-47
#>  4     4 /pine/scr… Aef1_SA… FBgn0005694… RCAACAAC… 2.53e-48   6.64e-46 4.04e-43
#>  5     5 /pine/scr… Aef1_Fl… FBgn0005694  CAACWACW  8.78e-48   3.05e-45 1.85e-42
#>  6     6 /pine/scr… jim_F1-… FBgn0027339… RMAAAAAA… 1.12e-45   1.01e-42 6.13e-40
#>  7     7 /pine/scr… rn_SOLE… FBgn0259172… VNNNRCAA… 4.49e-35   3.07e-32 1.87e-29
#>  8     8 /pine/scr… jim_SAN… FBgn0027339  AAAAAAAA… 5.37e-33   2.19e-30 1.33e-27
#>  9     9 /pine/scr… CG4360_… FBgn0038787  MCAACAAC… 7.06e-31   1.71e-28 1.04e-25
#> 10    10 /pine/scr… fru-PG_… FBgn0004652… GAAAAAW   2.66e-31   2.91e-28 1.77e-25
#> # … with 97 more rows, and 9 more variables: tests <int>, fasta_max <dbl>,
#> #   pos <int>, neg <int>, pwm_min <dbl>, tp <int>, tp_percent <dbl>, fp <int>,
#> #   fp_percent <dbl>