Import tomtom data from previous run

Import tomtom data from previous run

importTomTomXML(tomtom_xml_path)

Arguments

tomtom_xml_path
path to tomtom.xml

Value

will return data.frame with input motifs & results for best match. `tomtom` list column contains full tomtom data for each input motif. NOTE: if tomtom detects no matches for any input motif, currently will print a message & return NA values for `tomtom`, `best_match_name`, and `best_match_motif`.

Details

tomtom list column format the `tomtom` list column contains data.frames with the following format: - name: name of query PWM - altname: alternate name of query PWM - match_name: name of matched PWM - match_altname: alt name of matched PWM - match_pval: p-value of match - match_eval: E-value of match - match_qval: q-value of match - match_offset: number of letters the query was offset from the target match - match_strand: whether the motif was found on input strand (+) or as reverse-complement (-) - db_name: database source of matched motif - match_motif: universalmotif object containing the PWM that was matched

See also

Examples

tomtom_xml <- system.file("extdata", "tomtom.xml", package = "memes")
importTomTomXML(tomtom_xml)
#>       name altname family organism consensus alphabet strand icscore nsites
#> 1 AATAATAA DREME-2   <NA>     <NA>  AATAATAA      DNA     +-      16     10
#> 2      AAA DREME-1   <NA>     <NA>       AAA      DNA     +-       6      5
#> 3      AAT DREME-3   <NA>     <NA>       AAT      DNA     +-       6     10
#>   bkgsites pval qval eval best_match_name best_match_altname
#> 1       NA   NA   NA    9     bab1_FlyReg        FBgn0004870
#> 2       NA   NA   NA   39            <NA>               <NA>
#> 3       NA   NA   NA   12            <NA>               <NA>
#>                                                       best_match_motif
#> 1 <S4 class ‘universalmotif’ [package “universalmotif”] with 20 slots>
#> 2                                                                 NULL
#> 3                                                                 NULL
#>           best_db_name best_match_offset best_match_pval best_match_eval
#> 1 fly_factor_survey_id                 0        0.000287           0.174
#> 2                 <NA>                NA              NA              NA
#> 3                 <NA>                NA              NA              NA
#>   best_match_qval best_match_strand
#> 1           0.326                 -
#> 2              NA              <NA>
#> 3              NA              <NA>
#>                                                                  motif
#> 1 <S4 class ‘universalmotif’ [package “universalmotif”] with 20 slots>
#> 2 <S4 class ‘universalmotif’ [package “universalmotif”] with 20 slots>
#> 3 <S4 class ‘universalmotif’ [package “universalmotif”] with 20 slots>
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    tomtom
#> 1 bab1_FlyReg, ap_FlyReg, Pph13_Cell, jigr1_SANGER_5, So_Cell, Antp_FlyReg, br-Z4_FlyReg, hb_FlyReg, zen_FlyReg, Hgtx_SOLEXA, eve_FlyReg, FBgn0004870, FBgn0000099_2, FBgn0023489, FBgn0039350, FBgn0003460, FBgn0000095_2, FBgn0000210_4, FBgn0001180_2, FBgn0004053_2, FBgn0040318_2, FBgn0000606_2, <S4 class ‘universalmotif’ [package “universalmotif”] with 20 slots>, <S4 class ‘universalmotif’ [package “universalmotif”] with 20 slots>, <S4 class ‘universalmotif’ [package “universalmotif”] with 20 slots>, <S4 class ‘universalmotif’ [package “universalmotif”] with 20 slots>, <S4 class ‘universalmotif’ [package “universalmotif”] with 20 slots>, <S4 class ‘universalmotif’ [package “universalmotif”] with 20 slots>, <S4 class ‘universalmotif’ [package “universalmotif”] with 20 slots>, <S4 class ‘universalmotif’ [package “universalmotif”] with 20 slots>, <S4 class ‘universalmotif’ [package “universalmotif”] with 20 slots>, <S4 class ‘universalmotif’ [package “universalmotif”] with 20 slots>, <S4 class ‘universalmotif’ [package “universalmotif”] with 20 slots>, fly_factor_survey_id, fly_factor_survey_id, fly_factor_survey_id, fly_factor_survey_id, fly_factor_survey_id, fly_factor_survey_id, fly_factor_survey_id, fly_factor_survey_id, fly_factor_survey_id, fly_factor_survey_id, fly_factor_survey_id, 0, -2, 0, 0, 0, -1, 0, -1, 0, 0, 1, 0.000287, 0.00413, 0.00419, 0.00431, 0.00657, 0.00825, 0.0086, 0.0102, 0.012, 0.0145, 0.0164, 0.174, 2.51, 2.55, 2.62, 4, 5.02, 5.23, 6.18, 7.31, 8.84, 9.99, 0.326, 0.592, 0.592, 0.592, 0.592, 0.592, 0.592, 0.592, 0.592, 0.592, 0.592, -, +, +, +, +, +, +, -, -, +, -
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          NA, NA, NA, NA, NA, NA, NA, NA
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          NA, NA, NA, NA, NA, NA, NA, NA