Remove duplicated motif entries

This function identifies motif matrices which are duplicated in a universalmotif list or universalmotif_df and removes them. This operation ignores motif metadata and operates by removing all entries whose motif matrices are identical. The first instance of a duplicated motif in the input list is the one returned.

remove_duplicate_motifs(x)

Arguments

x
a universalmotif list or universalmotif_df

Value

A deduplicated list or universalmotif_df

Examples

motif <- universalmotif::create_motif()
duplicated <- c(motif, motif)
remove_duplicate_motifs(duplicated)

#> [[1]]
#> 
#>        Motif name:   motif
#>          Alphabet:   DNA
#>              Type:   PPM
#>           Strands:   +-
#>          Total IC:   10.79
#>       Pseudocount:   0
#>         Consensus:   MTSTNTCGGS
#> 
#>      M    T    S    T    N    T    C    G    G    S
#> A 0.51 0.15 0.00 0.00 0.31 0.17 0.04 0.00 0.00 0.01
#> C 0.46 0.06 0.50 0.00 0.00 0.15 0.62 0.00 0.00 0.36
#> G 0.02 0.10 0.48 0.01 0.44 0.00 0.19 0.97 0.99 0.62
#> T 0.01 0.69 0.02 0.99 0.25 0.68 0.14 0.02 0.01 0.00
#>