Run FIMO on input sequences

NOTE: runs best using text = TRUE, setting genomic coordinates in fasta headers, and using parse_genomic_coord = TRUE.

runFimo(
  sequences,
  motifs,
  bfile = "motif",
  outdir = "auto",
  parse_genomic_coord = TRUE,
  skip_matched_sequence = TRUE,
  max_strand = TRUE,
  text = TRUE,
  meme_path = NULL,
  silent = TRUE,
  ...
)

Arguments

sequences
path to fasta file, or stringset input.

motifs
path to .meme format file, or universalmotif/universalmotif list input.

bfile
path to background file, or special values: "motif", or "uniform" to use 0-order frequencies contained in the motif, or a uniform letter distribution. (default: "motif")

outdir
output directory location. Only used if text = FALSE. Default: "auto" to autogenerate directory name. Note: if not using a fasta path as input, this will be a temporary location unless explicity set.

parse_genomic_coord
logical(1) whether to parse genomic position from fasta name. Fasta names must be UCSC format positions (ie "chr:start-end"). If names of fasta entries are genomic coordinates and parse_genomic_coord == TRUE, results will contain positions of motif matches, otherwise FIMO will count positions from 1 to length of fasta file.

skip_matched_sequence
logical(1) whether or not to include the DNA sequence of the match. Default: FALSE. Note: jobs will complete faster if set to FALSE. add_sequence() can be used to lookup the sequence after data import if parse_genomic_coord is TRUE, so setting this flag is not strictly needed.

max_strand
if match is found on both strands, only report strand with best match.

text
logical(1) (default: TRUE). No output files will be created on the filesystem. The results are unsorted and no q-values are computed. This setting allows fast searches on very large inputs.

meme_path
path to meme/bin/ (optional). Defaut: NULL, searches "MEME_PATH" environment variable or "meme_path" option.

silent
logical(1) whether to suppress stdout/stderr printing to terminal (default: TRUE). NOTE: if text = TRUE, setting silent = TRUE will print all FIMO matches to terminal if the run is successful. Can be useful for troubleshooting.

...
additional commandline arguments to fimo see: http://meme-suite.org/doc/fimo.html?man_type=web

Value

GenomicRanges object contining positions of each match. Note: if parse_genomic_coords = FALSE, each seqnames entry will be the fasta header, and start/end will be the position within that sequence of the match. It is a good idea to use parse_genomic_coords = TRUE if using get_sequence() to create input sequences.

Details

Additional arguments passed to ...

FIMO Flag	allowed values	default	description

alpha	numeric(1)	1	alpha for calculating position-specific priors. Represents fraction of sites that are binding sites of TF of interest. Used in conjunction with psp

bfile	"motif", "motif-file", "uniform", file path,	"motif"	If "motif" or "motif-file", use 0-order letter frequencies from motif. "uniform" sets uniform letter frequencies.

max_stored_scores	integer(1)	NULL	maximum number of scores to be stored for computing q-values. used when text = FALSE, see FIMO webpage for details

motif_pseudo	numeric(1)	0.1	pseudocount added to motif matrix

no_qvalue	logical(1)	FALSE	only needed when text = FALSE, do not compute q-value for each p-value

norc	logical(1)	FALSE	Do not score reverse complement strand

prior_dist	file path	NULL	file containing binned distribution of priors

psp	file path	NULL	file containing position specific priors. Requires prior_dist

qv_thresh	logical(1)	FALSE	use q-values for the output threshold

thresh	numeric(1)	1e-4	output threshold for returning a match, only matches with values less than thresh are returned.

Licensing

The MEME Suite is free for non-profit use, but for-profit users should purchase a license. See the MEME Suite Copyright Page for details.

Citation

If you use runFimo() in your analysis, please cite:

Charles E. Grant, Timothy L. Bailey, and William Stafford Noble, "FIMO: Scanning for occurrences of a given motif", Bioinformatics, 27(7):1017-1018, 2011. full text

Examples

if (FALSE) {
seq <- universalmotif::create_sequences()
# sequences must have names in their fasta headers
names(seq) <- seq_along(seq)
motif <- universalmotif::create_motif()
runFimo(seq, motif, parse_genomic_coord = FALSE)
}