Identify motifs with MEME

NOTE: add note about the hsfrac issue

runMeme(
  input,
  control = NA,
  outdir = "auto",
  alph = "dna",
  parse_genomic_coord = TRUE,
  combined_sites = FALSE,
  silent = TRUE,
  meme_path = NULL,
  ...
)

Arguments

input
path to fasta, Biostrings::BStringSet list, or list of Biostrings::BStringSet (can generate using `get_sequence()`)

control
any data type as in `input`, or a character vector of `names(input)` to use those regions as control sequences. Using sequences as background requires an alternative objective function. Users must pass a non-default value of `objfun` to `...` if using a non-NA control set (default: NA)

outdir
(default: "auto")

alph
one of c("dna", "rna", "protein") or path to alphabet file (default: "dna").

parse_genomic_coord
`logical(1)` whether to parse genomic coordinates from fasta headers. Requires headers are in the form: "chr:start-end", or will result in an error. Automatically set to `FALSE` if `alph = "protein"`. This setting only needs to be changed if using a custom-built fasta file without genomic coordinates in the header.

combined_sites
`logical(1)` whether to return combined sites information (coerces output to list) (default: FALSE)

silent
Whether to suppress printing stdout to terminal (default: TRUE)

meme_path
path to "meme/bin/". If unset, will use default search behavior: 1. `meme_path` setting in `options()` 2. `MEME_PATH` setting in `.Renviron`

...

Value

Details

Additional arguments

`runMeme()` accepts all valid arguments to meme as arguments passed to `...`. For flags without values, pass them as `flag = TRUE`. The `dna`, `rna`, and `protein` flags should instead be passed to the `alph` argument of `runMeme()`. The arguments passed to MEME often have many interactions with eachother, for a detailed description of each argument see the [MEME Help Page](meme-suite.org/doc/meme.html).

For use with ChIP-seq data, the suggestion is to set: `revcomp = TRUE, minw = 5, maxw = 20`. See [using MEME with ChIP-Seq Data Tips](https://groups.google.com/forum/# for additional details.

# Citation

If you use `runMeme()` in your analysis, please cite:

Timothy L. Bailey and Charles Elkan, "Fitting a mixture model by expectation maximization to discover motifs in biopolymers", Proceedings of the Second International Conference on Intelligent Systems for Molecular Biology, pp. 28-36, AAAI Press, Menlo Park, California, 1994. [pdf](https://tlbailey.bitbucket.io/papers/ismb94.pdf)

## Licensing The MEME Suite is free for non-profit use, but for-profit users should purchase a license. See the [MEME Suite Copyright Page](http://meme-suite.org/doc/copyright.html) for details.