Update the `motif` column to data.frame values

Update the `motif` column to data.frame values

update_motifs(.data)

Arguments

.data
data.frame with `motif` column

Value

.data where `motif` column has been updated to reflect the values from columns sharing names with unprotected universalmotif slots. Names of `motif` list are updated to reflect name.

Details

**NOTE** that `consensus`, `alphabet`, `multifreq`, and `icscore` are protected columns and cannot be updated. Any changes to these columns will be ignored when updating motifs.

## Table of values updated

| `motif` | `data.frame` | |:--------:|:------------:| | name | name | | altname | altname | | family | family | | organism | organism | | strand | strand | | nsites | nsites | | bkgsites | bkgsites | | pval | pval | | qval | qval | | eval | eval |

Examples

motif <- universalmotif::create_motif()
df <- as_universalmotif_dataframe(motif)
df <- dplyr::mutate(df, id = "newName")
df <- update_motifs(df)
# renamed motif
df$motif

#> $motif
#> 
#>        Motif name:   motif
#>          Alphabet:   DNA
#>              Type:   PPM
#>           Strands:   +-
#>          Total IC:   11.99
#>       Pseudocount:   0
#>         Consensus:   SRYWTATGGH
#> 
#>      S    R    Y    W    T    A    T G G    H
#> A 0.00 0.67 0.05 0.50 0.00 0.77 0.06 0 0 0.45
#> C 0.34 0.00 0.59 0.15 0.00 0.22 0.00 0 0 0.26
#> G 0.66 0.33 0.07 0.05 0.04 0.01 0.00 1 1 0.01
#> T 0.01 0.00 0.28 0.30 0.96 0.00 0.94 0 0 0.28
#>