Import MEME results

This is a light wrapper around [universalmotif::read_meme()] that imports MEME results as universalmotif data.frame. If MEME is run with genomic coordinates in the fasta header, in "chr:start-end" format (base 1 indexed), the genomic coordinates of the motif match from input sequences can be parsed from the header.

importMeme(meme_txt, parse_genomic_coord = FALSE, combined_sites = FALSE)

Arguments

meme_txt
path to "meme.txt" output

parse_genomic_coord
whether to parse sequence headers into genomic coordinates for motif position information, only works if fasta files were written such that the sequence headers are in the form: "chr:start-end", or some variation of this form (delimiters can be any of: "[^[:alnum:]]+" (ie non-alphanumeric characters)), (default = FALSE).

combined_sites
whether to add `combined_sites` output which contains coordinates of each sequence, the motif sequence (if `parse_genomic_coord = TRUE`), and the `diagram` column raw output from MEME indicating the relative locations of motifs in the sequence.

Value

MEME results in universalmotif data.frame format (see: [as_universalmotif_dataframe()]). `sites_hits` is a nested data.frame column containing the position within each input sequence of matches to the identified motif.

See also

Examples

example_meme_txt <- system.file("extdata", "meme_full.txt", package = "universalmotif")
importMeme(example_meme_txt)
#>     motif name altname                     consensus alphabet strand  icscore
#> 1 <mot:1>    1    MEME               AAAAAAAAAARAAAS      DNA     +- 19.23570
#> 2 <mot:2>    2    MEME YNHTCNCNNYNTSACNCKTBYYYTYCNTY      DNA     +- 13.88457
#> 3 <mot:3>    3    MEME                   CAMMAMAAAAA      DNA     +- 14.81432
#>   nsites    eval type pseudocount                        bkg width
#> 1     50 5.3e-32  PPM           1 0.343, 0.157, 0.157, 0.343    15
#> 2     44 4.8e-07  PPM           1 0.343, 0.157, 0.157, 0.343    29
#> 3     48 7.6e+01  PPM           1 0.343, 0.157, 0.157, 0.343    11
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         sites_hits
#> 1                                                                                                                                                                                                                                                                                                                                                                       AT1G21910, AT5G57560, AT4G29780, AT4G29780, AT3G48650, AT2G25735, AT2G20670, AT2G20670, AT2G40000, AT1G80440, AT1G80440, AT1G27730, AT3G04640, AT1G56060, AT1G66090, AT1G25400, AT4G24570, AT1G24140, AT5G25240, AT1G56240, AT4G14450, AT4G08950, AT3G48650, AT1G80440, AT3G46090, AT5G56550, AT5G52050, AT3G48650, AT2G20670, AT2G20670, AT1G25400, AT3G04640, AT2G44840, AT5G25240, AT3G55980, AT5G61600, AT4G27654, AT3G55980, AT2G44840, AT1G66090, AT3G48650, AT2G40000, AT1G24140, AT1G66090, AT1G56240, AT3G55980, AT2G32200, AT2G40000, AT4G13395, AT4G08950, 744, 377, 510, 131, 951, 56, 455, 141, 853, 292, 911, 528, 766, 99, 197, 578, 762, 659, 30, 461, 572, 883, 166, 255, 363, 737, 640, 203, 356, 41, 630, 64, 313, 781, 52, 948, 48, 943, 873, 896, 5, 810, 622, 506, 123, 905, 343, 256, 889, 386, 5.51e-08, 1.35e-07, 1.35e-07, 1.35e-07, 1.35e-07, 1.35e-07, 1.35e-07, 1.35e-07, 1.35e-07, 1.35e-07, 2.35e-07, 2.45e-07, 3.77e-07, 7.55e-07, 8.88e-07, 8.88e-07, 1.22e-06, 1.22e-06, 1.39e-06, 1.39e-06, 1.58e-06, 1.75e-06, 1.75e-06, 1.75e-06, 1.99e-06, 2.13e-06, 2.13e-06, 2.13e-06, 2.36e-06, 2.36e-06, 2.9e-06, 3.15e-06, 3.15e-06, 3.62e-06, 3.92e-06, 4.35e-06, 4.35e-06, 4.35e-06, 4.35e-06, 4.35e-06, 4.64e-06, 4.64e-06, 5.39e-06, 5.87e-06, 5.87e-06, 6.72e-06, 7.08e-06, 7.65e-06, 8.74e-06, 1.15e-05, GAAAAAAAAAGAAAC, AAAAAAAAAAAAAAC, AAAAAAAAAAAAAAC, AAAAAAAAAAAAAAC, AAAAAAAAAAAAAAC, AAAAAAAAAAAAAAC, AAAAAAAAAAAAAAC, AAAAAAAAAAAAAAC, AAAAAAAAAAAAAAC, AAAAAAAAAAAAAAC, AAAAAAGAAAGAAAC, AAAAAAGAAAGAAAG, AAAAAAAAAAGAGAG, AAAAAAAAGAGCAAG, AAAAGAAAAAAAAAC, CAAGAAAAAAGAAAC, GAAAACAAAAGAAAC, AAAAAAAACAAAAAC, CAAAAAAAGAAAAAG, CAAAAAAAAAACAAG, CAAAACAAAAGAAAG, GAAAAAAAGAGAGAG, AAAAAAGAAAACAAC, AAAAAAAAGAACAAG, GAAAAAAACAGCAAC, CAAAAAAACAAAAAC, CAAGAAAAAAAAAAC, AAAAGAAGAAGAAAG, AAAAAAAAAAAAAAA, AAAAAAAAAAAAAAA, GAAAGAAGAAGAAAC, AAAAATAAAAGAAAG, AAAAAAGAGAGAGAG, AAAGAGAAAAGAAAG, CAAAAAAAAAAAAAA, AAAAAACAAAGAGAG, AAAAGAAACAAAAAG, AAAAAAAAGACCAAG, AAAAAAAGAAGAAAA, GAAAAAAGAACAAAC, CAAAAAAAAAGGAAG, AAAAACGAAAAAAAG, AAAAAGAAAAAAGAC, AAAAATAAAAAAAAG, AAAAGAAAAACCAAG, GAAGAACAAAGAAAG, AAAAGAAAAAGCAAA, GAAAAAGAGAGAAAA, GAAGAAGAAACAAAC, GAAAAGAGAAGAGAG
#> 2 AT4G29780, AT5G25240, AT4G25490, AT3G55980, AT2G40000, AT1G74930, AT4G24570, AT4G25490, AT1G76650, AT1G74930, AT1G35140, AT4G32480, AT4G13395, AT3G44260, AT1G25400, AT5G45630, AT4G25490, AT4G25490, AT2G32200, AT2G20670, AT3G55980, AT2G44130, AT1G74930, AT3G15500, AT5G45630, AT1G72910, AT1G80840, AT1G21910, AT4G14450, AT1G27730, AT1G56240, AT4G32480, AT4G25490, AT3G15500, AT1G27730, AT1G72910, AT5G56550, AT1G35140, AT5G57560, AT4G27270, AT5G25240, AT1G35210, AT1G32920, AT1G80840, 887, 805, 149, 128, 924, 517, 846, 827, 159, 937, 494, 815, 939, 884, 702, 280, 921, 71, 2, 880, 804, 936, 53, 409, 322, 273, 813, 925, 959, 920, 39, 899, 210, 488, 746, 806, 115, 579, 880, 178, 856, 935, 140, 916, 1.47e-12, 2.43e-09, 2.43e-09, 1.02e-08, 2.73e-08, 4.37e-08, 7.76e-08, 8.68e-08, 8.68e-08, 1.21e-07, 1.35e-07, 1.68e-07, 1.88e-07, 1.88e-07, 2.09e-07, 2.87e-07, 4.35e-07, 4.81e-07, 4.81e-07, 5.33e-07, 5.9e-07, 5.9e-07, 5.9e-07, 6.52e-07, 7.96e-07, 7.96e-07, 7.96e-07, 9.69e-07, 1.18e-06, 1.18e-06, 1.3e-06, 1.43e-06, 1.89e-06, 1.89e-06, 1.89e-06, 2.74e-06, 3.01e-06, 3.01e-06, 3.29e-06, 3.6e-06, 4.31e-06, 4.71e-06, 6.13e-06, 1.21e-05, CGTTCACACCTTCACGCGCTTCCTTCGTC, CGCTCCCGCCTTCACAGCTGTCTTTCATC, TCTTCACCTTACCACTCTTTTTTTCCCTC, CTCACACAACATGACTCTCCTCCTTGGTT, CACACGCCTTTTTTCTCGTCCCCATCCTC, TGATCAACAGACGACTCGCTTCTTTCTTC, CTGTCTCACTCACACACGCGTTTTCCTAC, CACTCGCTGTCTGACGTGTGTCCTTTGAC, CATTCACTTTCTCACCCTTGCTCTCTTCT, TCTTCCTCACTTGCCACTTCCACTTCTTC, TCTTCGCCACATAAATCTTGCACTTTCTC, CAATCATATACTGACACGTCATCTTCTTC, CGCGTCCGACATCACACGCTCTCTTCTAT, ACCCCACCACTTGACGCGTTTCTCGTCCC, TCGTCACCTCCTCAGTGATTCCTCCGATC, CATTCTCAACGTCACGCGTGTACGCGGTT, TGCTCCCCACATATCCGTTATCTCTCCTC, CCCACCAGCCTTCACTGGCCCCATACGTC, CCATCTTCACTTGTCTCTTTTCCATCATG, CAATCACAGTTTCACACGTCACCATTCGC, CGCTCTTGTGCTGACTCTTCTACTCGTAT, CAACCGCAAGCAAACTCATCCTCCTCTTT, CGACCACAACAACAAGAATCTTCTCCTTC, CTACCACAAGCCGACACATGTCTTGTGTC, GCGTTTCCTCCTGACGCGCCGCTCTGTTT, CATCCACATCCTCTCTCACTTTCTTTATG, TCAACTCGGCTTCTATCGGGTCCACCCGC, TCATCTCTTCCTCTCCCATACCCCTCCCA, TCAACATGTTACAAGTCGTGGCCCTCCTC, CCTACCACTTGTCACGCAACTTCCTTCTC, CAGTTCACTAATCACACTTTCTCTTGTTC, TCTCCCACCTCTCAGGCGGGTCCCTAACC, CTCTCTCGTTACCTCTCGCTCTCACGTTA, CCATCATTCACTTACACGTGTTTTTTCTT, CTTGCACACACACACACGTGTACTAGGTC, CTTTGGCGGCCCATCTATTCTCTTCCCTC, TCCTTAAACCAAGAATCTTGTCCCTCTTT, CGAACATTACATGACTTTCTTCTTCCCAC, TGTTGCCTCAAAGACGCGGCTTCTTCTTT, CAACCCAATTTCCACGATTTTCTTCCTAC, CGCTCCTCTCACATCACTGTCACACTCTT, CCATCCACGACCAAATCATTCTCTCCATT, TCGACCCTCCAACTCCATTGCCTATCTAC, GCTCCTCTCTCTCTCTTCTGCCTCTTTCC
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                AT4G13395, AT3G46090, AT1G66090, AT1G80440, AT4G24570, AT1G24140, AT5G54490, AT4G24570, AT1G35140, AT1G56240, AT1G69490, AT2G32200, AT5G25240, AT5G35777, AT4G32480, AT4G14450, AT3G15500, AT3G46090, AT1G76650, AT1G27730, AT2G44840, AT1G80440, AT4G27654, AT2G25735, AT1G27730, AT5G45630, AT4G27270, AT4G29780, AT3G50050, AT4G25490, AT1G56240, AT1G80840, AT1G80840, AT3G50050, AT1G33760, AT5G57560, AT5G25240, AT3G15500, AT3G55980, AT2G44840, AT5G51190, AT4G27270, AT3G48650, AT3G04640, AT1G66090, AT5G45630, AT1G21910, AT4G14450, 693, 611, 58, 524, 351, 150, 986, 983, 987, 893, 286, 442, 301, 641, 183, 853, 756, 659, 899, 811, 719, 892, 34, 289, 244, 190, 422, 989, 145, 296, 710, 354, 187, 747, 385, 402, 544, 525, 765, 804, 986, 761, 550, 3, 794, 174, 499, 467, 2.36e-06, 2.36e-06, 2.36e-06, 2.36e-06, 3.44e-06, 3.44e-06, 4.52e-06, 4.52e-06, 4.52e-06, 6.1e-06, 6.1e-06, 7.67e-06, 1.33e-05, 1.33e-05, 1.33e-05, 1.33e-05, 1.33e-05, 1.33e-05, 1.33e-05, 1.38e-05, 1.64e-05, 1.64e-05, 1.93e-05, 1.93e-05, 2e-05, 2.34e-05, 2.34e-05, 2.34e-05, 2.34e-05, 2.6e-05, 2.6e-05, 2.6e-05, 2.6e-05, 2.71e-05, 2.71e-05, 3.08e-05, 3.08e-05, 3.08e-05, 3.08e-05, 3.08e-05, 3.18e-05, 3.94e-05, 3.94e-05, 4.25e-05, 4.25e-05, 5.18e-05, 5.18e-05, 7.17e-05, CACAAAAAAAA, CACAAAAAAAA, CACAAAAAAAA, CACAACAAAAA, CACCAAAAAAA, CACCAAAAAAA, CACAACACAAA, CACAAAACAAA, CACAACACAAA, CACCAAACAAA, CACAACCAAAA, CCCAAAAAAAA, CAAAAAAAAAA, CAAAACAAAAA, CAAAAAAAAAA, CAAAACAAAAA, CAAAAAAAAAA, CAAAAAAAAAA, CAAAAAAAAAA, CCCCAAAAAAA, CAACAAAAAAA, CAACACAAAAA, CAAAAAACAAA, CAAAAAACAAA, CCCAACCAAAA, CAAAACCAAAA, CAACACACAAA, CAAAACCAAAA, CAAAAACAAAA, GACAAAAAAAA, GACAAAAAAAA, GACAAAAAAAA, GACAAAAAAAA, CAACAACAAAA, CAACAACAAAA, GACCACAAAAA, GACCAAAAAAA, CCAAAAAAAAA, GACCAAAAAAA, CCAAACAAAAA, CAAAAACCAAA, CACACAAAAAA, CCAAAAACAAA, CACCCCAAAAA, CCAAAACAAAA, CACAACACAAC, CACAAAACAAC, CACGAACAAAA
#> 
#> [Hidden empty columns: family, organism, bkgsites, pval, qval.]